+++ torch_geometric/data/download.py
@@ -1,3 +1,4 @@
+import os
 import os.path as osp
 import ssl
 import sys
@@ -6,7 +7,6 @@ from typing import Optional
 
 import fsspec
 
-from torch_geometric.data.makedirs import makedirs
 from torch_geometric.io import fs
 
 
@@ -41,7 +41,7 @@ def download_url(
     if log and 'pytest' not in sys.modules:
         print(f'Downloading {url}', file=sys.stderr)
 
-    makedirs(folder)
+    os.makedirs(folder, exist_ok=True)
 
     context = ssl._create_unverified_context()
     data = urllib.request.urlopen(url, context=context)